Add vigenere cipher operation
This commit adds the Vigenère cipher operation to the
@ciphereditor/extension-essentials.  The code is based on the Caesar
cipher and cryptii code for the Vigenère cipher.

Four different variants of the cipher are available, just like Cryptii.
1. Traditional Vigenère cipher: inputIndex + alphabetIndex,
2. Beaufort cipher: alphabetIndex - inputIndex,
3. Variant Beaufort cipher: inputIndex - alphabetIndex,
4. Trithemius cipher: Vigenere with fixed key.

The code gives an error when the key is empty or contains characters not
in the alphabet.  If none of the plain-/ciphertext characters are from
the alphabet, it warns the user to avoid confusion.

The Trithemius cipher uses a fixed key and key mode.  Currently, this is
not communicated to the user, this can be part of an upcoming update.

Signed-off-by: anderium <33520919+anderium@users.noreply.github.com>